+++ src/__test__/utils/graphVisualisation/setNodesIdsToDisplay.test.js
@@ -4,7 +4,8 @@ import {
   ALGO_TYPE_SHORTEST_PATH,
   ALGO_TYPE_BOUNDING_BOX,
   ALGO_TYPE_NODES_FILTER,
-  ALGO_TYPE_EDGES_FILTER
+  ALGO_TYPE_EDGES_FILTER,
+  ALGO_TYPE_SEARCH_NEIGHBOURHOOD
 } from '../../../constants/algorithms'
 import setNodesIdsToDisplay from '../../../utils/graphVisualisation/setNodesIdsToDisplay'
 import { classesFromApi } from '../../fixtures/classesFromApi'
@@ -48,6 +49,10 @@ describe('setNodesIdsToDisplay', () => {
       [
         [
           'highlightedNodes',
+          [],
+        ],
+        [
+          'highlightedEdges',
           [],
         ],
         [
@@ -286,6 +291,10 @@ describe('setNodesIdsToDisplay', () => {
         [],
       ],
       [
+        'highlightedEdges',
+        [],
+      ],
+      [
         'isNodeOverlay',
         false,
       ],
@@ -303,6 +312,55 @@ describe('setNodesIdsToDisplay', () => {
           '12',
           '24',
         ],
+      ],
+    ])
+  })
+
+  it('should work correctly when ALGO_TYPE_SEARCH_NEIGHBOURHOOD', async () => {
+    const type = ALGO_TYPE_SEARCH_NEIGHBOURHOOD
+
+    const options = {
+      selectedNodesId: ['12', '24'],
+      separationDegree: 1,
+      selectedEdgesId: ['111'],
+      edgesPerNode
+    }
+
+    getNeighbours.mockImplementation(() => ['24', '36'])
+
+    await setNodesIdsToDisplay({
+      type,
+      setStoreState,
+      options
+    })
+
+    expect(setStoreState.mock.calls).toEqual([
+      [
+        'highlightedNodes',
+        ['12', '24'],
+      ],
+      [
+        'highlightedEdges',
+        ['111'],
+      ],
+      [
+        'isNodeOverlay',
+        false,
+      ],
+      [
+        'shortestPathNodes',
+        [],
+      ],
+      [
+        'shortestPathResults',
+        [],
+      ],
+      [
+        'nodesIdsToDisplay',
+        [
+          '24',
+          '36',
+        ]
       ],
     ])
   })
@@ -330,6 +388,10 @@ describe('setNodesIdsToDisplay', () => {
     expect(setStoreState.mock.calls).toEqual([
       [
         'highlightedNodes',
+        ['12'],
+      ],
+      [
+        'highlightedEdges',
         [],
       ],
       [
@@ -390,6 +452,10 @@ describe('setNodesIdsToDisplay', () => {
           '24',
           '48',
         ],
+      ],
+      [
+        'highlightedEdges',
+        [],
       ],
       [
         'isNodeOverlay',
@@ -457,6 +523,10 @@ describe('setNodesIdsToDisplay', () => {
         ],
       ],
       [
+        'highlightedEdges',
+        [],
+      ],
+      [
         'isNodeOverlay',
         false,
       ],
@@ -514,6 +584,10 @@ describe('setNodesIdsToDisplay', () => {
         [],
       ],
       [
+        'highlightedEdges',
+        [],
+      ],
+      [
         'isNodeOverlay',
         false,
       ],
@@ -556,6 +630,10 @@ describe('setNodesIdsToDisplay', () => {
     expect(setStoreState.mock.calls).toEqual([
       [
         'highlightedNodes',
+        [],
+      ],
+      [
+        'highlightedEdges',
         [],
       ],
       [