+++ src/__test__/utils/nodeNeighbourhood/getNeighbours.test.js
@@ -17,7 +17,8 @@ const edges = Object.keys(objectPropertiesFromApi).map((property) => ({
 store.getState = jest.fn().mockImplementation(() => ({
   edgesPerNode,
   availableEdges: new DataSet(edges),
-  deletedNodes
+  deletedNodes,
+  objectPropertiesFromApi
 }))
 
 describe('getNeighbours', () => {