+++ src/utils/networkStyling/setEdgesStyle.js
@@ -1,5 +1,6 @@
 import resetEdgesStyles from './resetEdgesStyles'
 import setEdgesStylesByProperty from './setEdgesStylesByProperty'
+import setHighlightedEdges from './setHighlightedEdges'
 
 /**
  * Update edges style
@@ -9,6 +10,8 @@ const setEdgesStyle = () => {
   resetEdgesStyles()
 
   setEdgesStylesByProperty()
+
+  setHighlightedEdges()
 }
 
 export default setEdgesStyle